Parse github auth token from hub config file
I use the GitHub API to power a few of the git scripts I have, and previously I
relied on a token file associated with the gist-vim project. I'm not currently
using it, but I am using and expect to continue to use the hub CLI, so I can
borrow its token.