vim-git: Pre-squash when editing interactive rebase
This has some cryptic noise in the substitute command (`silent! w!` etc).
Originally I was just running the substitute command, but this seemed to create
a situation in which Vim (git?... both?) didn't actually recognize the file as
having been edited, and thus git would just cancel the rebase.

With that configuration, I had to manually edit the buffer in some way. Deleting
and undoing for a single character worked, but this obviously was a sub-par
solution.

I was able to automate this with the addition of the `silent! w!` noise which
instructs Vim to force write as part of the autocmd. Hooray! (thanks @joelq for
help with this).