vim: Update ale lint config to be more chill
The default configuration of Ale is to lint on all changes. While it's
neat to get non-blocking feedback as you work in a file (without having
to save), the experience can be a bit distracting as you're typing out
some changes with the lint warnings popping in and out as you're typing.

With this configuration, instead of linting on text change events, we
lint when entering / leaving insert mode, and when at rest (CursorHold
event with 1000ms timeout).